count with usize by default

+++ minimum_redundancy/src/lib.rs
@@ -72,7 +72,7 @@ impl<ValueType, D: TreeDegree> Coding<ValueType, D> {
     pub fn from_iter<Iter>(degree: D, iter: Iter) -> Self
         where Iter: IntoIterator, Iter::Item: Borrow<ValueType>, ValueType: Hash + Eq + Clone
     {
-        Self::from_frequencies(degree, HashMap::<ValueType, u32>::with_occurrences_of(iter))
+        Self::from_frequencies(degree, HashMap::<ValueType, usize>::with_occurrences_of(iter))
     }
 
     /// Returns total (summarized) number of code fragments of all values.
